refactor(prerender): extract submitForScraping into its own module
Splits Step 2 (submitForScraping) and its private helpers out of the 1600+
line handler.js into src/prerender/submit-for-scraping.js, following the
same pattern as path-suggestions/. It is not re-exported from handler.js —
handler.js only imports it internally to wire the AuditBuilder step; tests
import it directly from the new module. findPrerenderOpportunity and a new
getActiveSuggestionStats helper move to utils/utils.js, shared with
handler.js's detectWrongEdgeDeployedStatus.

Also redesigns the LLMO-6533/LLMO-6638 suggestion cap: instead of an early
gate that blocked all new URL submission once a domain hit 4000 active
suggestions, the cap is now applied as a final filter on the automatic
daily batch only. URLs that already have an active suggestion keep being
resubmitted for re-verification (so the count can naturally fall via
OUTDATED/FIXED transitions); only brand-new candidate URLs are dropped.
CSV and Slack-triggered runs are explicit operator actions and are never
subject to this filter.

Moves the submitForScraping test surface out of handler.test.js into a
dedicated submit-for-scraping.test.js, and updates the affected behaviour
test files' import/esmock targets accordingly.
